feat: build LinksTableComponent
use mat-table to display links.

create Link class with its Firestore converter.
fetch links from firestore and display with the table.
get rid of serverTimestamp and use Date when creating links.
did this because of null errors with converter at creation and fetching.
And also because the client side's time works too in this case.

add id of link to the link class.
edit and deploy cloud function to add id of created link to its data.
did this to edit links with editor (in a later commit).

add matToolTip for showing "Edit Link" on hover of table button.
add spinner to show when linksTable is loading links.