feat: setup authentication
create AuthComponent and make it show in /auth route in admin.
show signInWithGoogle or signOut button depending on Auth State.
after signing in, accept only signed in user that has admin customClaims,
else delete the just signed in user and sign them out.

move editor and linksTable access to a new DashboardComponent for the router's sake.
provide sidenav in AppComponent to Dashboard (for opening Editor),
through a new SidenavService.
show edit button in linksTable only if Obum is signed in.
update Firestore Rules to permit links editing if only Obum is signed in.